+++ cg/services/deliver_files/delivery_file_fetcher_service/raw_data_delivery_file_fetcher.py
@@ -10,6 +10,7 @@ from cg.services.deliver_files.delivery_file_fetcher_service.delivery_file_fetch
 from cg.services.deliver_files.delivery_file_fetcher_service.error_handling import (
     handle_missing_bundle_errors,
 )
+from cg.services.deliver_files.delivery_file_fetcher_service.exc import NoDeliveryFilesError
 from cg.services.deliver_files.delivery_file_fetcher_service.models import (
     DeliveryFiles,
     DeliveryMetaData,
@@ -64,6 +65,16 @@ class RawDataDeliveryFileFetcher(FetchDeliveryFilesService):
                 sample_files=raw_data_files,
             )
         )
+
+    @staticmethod
+    def _validate_delivery_has_content(delivery_files: DeliveryFiles) -> DeliveryFiles:
+        """Check if the delivery files has files to deliver."""
+        if delivery_files.sample_files:
+            return delivery_files
+        LOG.info(
+            f"No files to deliver for case {delivery_files.delivery_data.case_id} in ticket: {delivery_files.delivery_data.ticket_id}"
+        )
+        raise NoDeliveryFilesError
 
     @handle_missing_bundle_errors
     def _get_raw_data_files_for_sample(self, case_id: str, sample_id: str) -> list[SampleFile]:
